moving limits migrations for 1.5.9
