+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/postgres/v160/Migration.java
@@ -1,6 +1,5 @@
 package org.openmetadata.service.migration.postgres.v160;
 
-import static org.openmetadata.service.migration.utils.v160.MigrationUtil.addAppExtensionName;
 import static org.openmetadata.service.migration.utils.v160.MigrationUtil.addViewAllRuleToOrgPolicy;
 import static org.openmetadata.service.migration.utils.v160.MigrationUtil.migrateServiceTypesAndConnections;
 
@@ -17,7 +16,6 @@ public class Migration extends MigrationProcessImpl {
   @Override
   @SneakyThrows
   public void runDataMigration() {
-    addAppExtensionName(handle, collectionDAO, authenticationConfiguration, true);
     migrateServiceTypesAndConnections(handle, true);
     addViewAllRuleToOrgPolicy(collectionDAO);
   }
